Eitanya/bump envoy 1.26.x (#244)
* [WIP] Ambient dev

* update TestStore to use rootStore

* Bump envoy to 2023-02-22 a26fbd70eb42287c31ef8ccdd7b2aca96748cafa

* change log

* change log

* Bump envoy to v1.26-dev 17844b10a3e57e411f3b11c595445327d6ac4d49

* Removed flaky test reporting upstream

https://github.com/envoyproxy/envoy/commit/92f1cd4685191ef26409230999a40fbca96f5247

* need VERSION.txt now

* fix ASSERT

* python

* tests pass locally

* update bazelversion

* I'm confused

* fix the changelog

* sub-folder

* more scripts

* update image

* use a specific hash

* remove upstream files

* Update ci/do_ci.sh

* Update ci/do_ci.sh

---------

+++ ci/do_ci.sh
@@ -4,17 +4,28 @@ bazel fetch //source/exe:envoy-static
 
 SOURCE_DIR="$(bazel info workspace)"
 
+# will be reverted or updated in https://github.com/solo-io/envoy-gloo/issues/246
+git clone https://github.com/envoyproxy/envoy.git /tmp/envoy
+pushd /tmp/envoy
+git checkout v1.25.7
+popd
+
 $SOURCE_DIR/ci/verify_posture.sh verify
 
-export UPSTREAM_ENVOY_SRCDIR=$(bazel info output_base)/external/envoy
+export UPSTREAM_ENVOY_SRCDIR=/tmp/envoy
 cp -f $UPSTREAM_ENVOY_SRCDIR/.bazelrc $SOURCE_DIR/
 # dont think this is needed... cp -f $UPSTREAM_ENVOY_SRCDIR/*.bazelrc $SOURCE_DIR/
 cp -f $UPSTREAM_ENVOY_SRCDIR/.bazelversion $SOURCE_DIR/.bazelversion
 # cp -f $UPSTREAM_ENVOY_SRCDIR/bazel/get_workspace_status $SOURCE_DIR/bazel/get_workspace_status
 cp -f $UPSTREAM_ENVOY_SRCDIR/ci/WORKSPACE.filter.example $SOURCE_DIR/ci/
+cp -f $UPSTREAM_ENVOY_SRCDIR/VERSION.txt $SOURCE_DIR/VERSION.txt
 
+# upstream removed the flaky_test
 mkdir -p $SOURCE_DIR/ci/flaky_test
-cp -a $UPSTREAM_ENVOY_SRCDIR/ci/flaky_test $SOURCE_DIR/ci
+# These were removed upstream but the build seems to fail here.
+if [[ -d "$UPSTREAM_ENVOY_SRCDIR/ci/flaky_test" ]] || [[ -f "$UPSTREAM_ENVOY_SRCDIR/ci/flaky_test" ]]; then
+  cp -a $UPSTREAM_ENVOY_SRCDIR/ci/flaky_test $SOURCE_DIR/ci
+fi
 
 cp -f $UPSTREAM_ENVOY_SRCDIR/tools/shell_utils.sh $SOURCE_DIR/tools
